pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Anton Arkhipov <antonolegovich44@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
